ECOPIC-main/middleware/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-dotenv==1.0.0
redis==4.6.0
celery==5.3.1
gunicorn==21.2.0